+++ MacPacker/Core/ArchiveWindow.swift
@@ -12,6 +12,9 @@ class ArchiveWindowController: NSWindowController, NSWindowDelegate {
     let archiveState: ArchiveState
     let appDelegate: AppDelegate
     
+    let archiveService = ArchiveService()
+    let contentService: ArchiveContentService = ArchiveContentService()
+    
     var willCloseHandler: (() -> Void)?
     
     init(archiveState: ArchiveState, appDelegate: AppDelegate) {
@@ -25,9 +28,23 @@ class ArchiveWindowController: NSWindowController, NSWindowDelegate {
         
         window.delegate = self
         
+        if #unavailable(macOS 14) {
+            // we have to use a classic toolbar for macOS 13 based on NSToolbar
+            // > no need for an else here as SwiftUIs .toolbar() loads the
+            //   toolbar in ContentView
+            let toolbar = NSToolbar(identifier: "ArchiveWindowToolbar")
+            toolbar.delegate = self
+            toolbar.displayMode = .iconOnly
+            window.toolbar = toolbar
+            
+            window.title = Bundle.main.appName
+        }
+        
+        window.toolbarStyle = .unified
+        
         // show the content view
         let contentView = ContentView()
-            .environment(AppState.shared)
+            .environmentObject(AppState.shared)
             .environmentObject(appDelegate)
             .environmentObject(archiveState)