+++ MacPacker/Model/AppState.swift
@@ -6,11 +6,10 @@
 //
 
 import Foundation
-import Observation
+import KeyboardShortcuts
 import SwiftUI
 
-@Observable
-class AppState {
+class AppState: ObservableObject {
     static var shared: AppState = AppState()
     
     private init() { }
@@ -21,4 +20,22 @@ class AppState {
     
     /// Holds all archives that the app currently takes care of
     var archives: [UUID: Archive2] = [:]
+}
+
+extension KeyboardShortcuts.Name {
+    static let spacePreview = Self("spacePreview", default: .init(.space, modifiers: []))
+}
+
+extension AppState {
+    func setUpEvents() {
+        KeyboardShortcuts.onKeyUp(for: .spacePreview) {
+            print("Space Preview")
+//            if let archive = state.archive,
+//               let selectedItem = state.selectedItems.first,
+//               let url = archive.extractFileToTemp(selectedItem) {
+//                appDelegate.openPreviewerWindow(for: url)
+//            }
+//            return .handled
+        }
+    }
 }